# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill. The config scope also
    grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/provider-sources:
    get:
      tags: [Provider Sources]
      summary: List provider sources
      operationId: listProviderSources
      x-astrbot-scope: provider
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Provider Sources]
      summary: Create a provider source
      operationId: createProviderSource
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderSourceConfigRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/provider-sources/by-id:
    get:
      tags: [Provider Sources]
      summary: Get a provider source by query ID
      operationId: getProviderSourceById
      x-astrbot-scope: provider
      parameters:
        - name: source_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Provider Sources]
      summary: Update a provider source by body ID
      operationId: upsertProviderSourceById
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_id, config]
              properties:
                source_id:
                  type: string
                config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Provider Sources]
      summary: Delete a provider source by query ID
      operationId: deleteProviderSourceById
      x-astrbot-scope: provider
      parameters:
        - name: source_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/provider-sources/models:
    get:
      tags: [Provider Sources]
      summary: List models available from a provider source by query ID
      operationId: listProviderSourceModelsById
      x-astrbot-scope: provider
      parameters:
        - name: source_id
          in: query
          required: true
          schema:
            type: string
        - name: capability
          in: query
          schema:
            $ref: "#/components/schemas/ProviderCapability"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/provider-sources/providers:
    get:
      tags: [Providers]
      summary: List providers under a provider source by query ID
      operationId: listProvidersBySourceId
      x-astrbot-scope: provider
      parameters:
        - name: source_id
          in: query
          required: true
          schema:
            type: string
        - name: capability
          in: query
          schema:
            $ref: "#/components/schemas/ProviderCapability"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Providers]
      summary: Create a provider under a provider source by body ID
      operationId: createProviderInSourceById
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_id, config]
              properties:
                source_id:
                  type: string
                config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/provider-sources/{source_id}:
    get:
      tags: [Provider Sources]
      summary: Get a provider source
      operationId: getProviderSource
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Provider Sources]
      summary: Update a provider source or create it if missing
      operationId: upsertProviderSource
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/SourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderSourceConfigRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Provider Sources]
      summary: Delete a provider source and providers that reference it
      operationId: deleteProviderSource
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/provider-sources/{source_id}/models:
    get:
      tags: [Provider Sources]
      summary: List models available from a provider source
      operationId: listProviderSourceModels
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: capability
          in: query
          schema:
            $ref: "#/components/schemas/ProviderCapability"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/provider-sources/{source_id}/providers:
    get:
      tags: [Providers]
      summary: List providers under a provider source
      operationId: listProvidersBySource
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: capability
          in: query
          schema:
            $ref: "#/components/schemas/ProviderCapability"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Providers]
      summary: Create a provider under a provider source
      operationId: createProviderInSource
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/SourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderConfigRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers:
    get:
      tags: [Providers]
      summary: List providers by capability or source
      operationId: listProviders
      x-astrbot-scope: provider
      parameters:
        - name: capability
          in: query
          schema:
            $ref: "#/components/schemas/ProviderCapability"
        - name: source_id
          in: query
          schema:
            type: string
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Providers]
      summary: Create a standalone provider
      operationId: createProvider
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderConfigRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/by-id:
    get:
      tags: [Providers]
      summary: Get a provider configuration by query ID
      operationId: getProviderById
      x-astrbot-scope: provider
      parameters:
        - name: provider_id
          in: query
          required: true
          schema:
            type: string
        - name: merged
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Providers]
      summary: Replace a provider configuration by body ID
      operationId: updateProviderById
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id, config]
              properties:
                provider_id:
                  type: string
                config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Providers]
      summary: Delete a provider by query ID
      operationId: deleteProviderById
      x-astrbot-scope: provider
      parameters:
        - name: provider_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/enabled:
    patch:
      tags: [Providers]
      summary: Enable or disable a provider by body ID
      operationId: setProviderEnabledById
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id, enabled]
              properties:
                provider_id:
                  type: string
                enabled:
                  type: boolean
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/test:
    post:
      tags: [Providers]
      summary: Test a provider by body ID
      operationId: testProviderById
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id]
              properties:
                provider_id:
                  type: string
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/embedding-dimension:
    post:
      tags: [Providers]
      summary: Probe embedding dimension for a provider by body ID
      operationId: getProviderEmbeddingDimensionById
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id]
              properties:
                provider_id:
                  type: string
                provider_config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/agy-cli/status:
    post:
      tags: [Providers]
      summary: Detect the managed or external Agy CLI installation
      operationId: getAgyCliStatus
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgyCLIActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/agy-cli/install:
    post:
      tags: [Providers]
      summary: Install or update the managed Agy CLI binary
      operationId: installOrUpdateAgyCli
      x-astrbot-scope: provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgyCLIActionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/{provider_id}:
    get:
      tags: [Providers]
      summary: Get a provider configuration
      operationId: getProvider
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - name: merged
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Providers]
      summary: Replace a provider configuration
      operationId: updateProvider
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderConfigRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Providers]
      summary: Delete a provider
      operationId: deleteProvider
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/{provider_id}/enabled:
    patch:
      tags: [Providers]
      summary: Enable or disable a provider
      operationId: setProviderEnabled
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnabledPatch"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/providers/{provider_id}/test:
    post:
      tags: [Providers]
      summary: Test a provider
      operationId: testProvider
      x-astrbot-scope: provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    SourceId:
      name: source_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    ProviderCapability:
      type: string
      enum:
        - chat
        - agent
        - stt
        - tts
        - embedding
        - rerank
...
    ProviderSourceConfigRequest:
      type: object
      required: [config]
      properties:
        id:
          type: string
        config:
          $ref: "#/components/schemas/DynamicConfig"
      additionalProperties: false
    ProviderConfigRequest:
      type: object
      required: [config]
      properties:
        id:
          type: string
        provider_source_id:
          type: string
        capability:
          $ref: "#/components/schemas/ProviderCapability"
        enabled:
          type: boolean
          default: true
        config:
          $ref: "#/components/schemas/DynamicConfig"
      additionalProperties: false
...